+++ components/tenders.tsx
@@ -5,10 +5,10 @@ import { motion, AnimatePresence } from "framer-motion"
 import { Search, Filter, Loader2, FileText, Info } from "lucide-react"
 import { Input } from "@/components/ui/input"
 import { Button } from "@/components/ui/button"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Badge } from "@/components/ui/badge"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/common/select"
+import { Badge } from "@/components/common/badge"
 import { TenderCard } from "@/components/tenders/tender-card"
-import { TenderDetailModal } from "@/components/tender-detail-modal"
+import TenderDetailModal from "@/components/tender-detail-modal"
 import { Tender } from "@/types/tender"
 import { axiosInstance } from "@/lib/axios"
